Improvements following review of WORD_BYTES=16 pull request
