+++ src/include/mit/mit.h.in
@@ -24,83 +24,61 @@
 #define ENDISM @ENDISM@
 #define HOST_ENDISM @WORDS_BIGENDIAN@
 
-// Types and printf formats
-typedef uint8_t mit_BYTE;
+// Types and constants
 #if WORD_BYTES == 2
-typedef int16_t mit_WORD;
-typedef uint16_t mit_UWORD;
+#define mit_WORD_BIT 16
 #define mit_SIZE_WORD 1
-#define mit_WORD_MASK UINT16_MAX
-#define mit_UWORD_MAX UINT16_MAX
-#define mit_WORD_MIN INT16_MIN
-#define mit_WORD_MAX INT16_MAX
-#define reverse_endianness bswap_16
 #elif WORD_BYTES == 4
-typedef int32_t mit_WORD;
-typedef uint32_t mit_UWORD;
+#define mit_WORD_BIT 32
 #define mit_SIZE_WORD 2
-#define mit_WORD_MASK UINT32_MAX
-#define mit_UWORD_MAX UINT32_MAX
-#define mit_WORD_MIN INT32_MIN
-#define mit_WORD_MAX INT32_MAX
-#define reverse_endianness bswap_32
 #elif WORD_BYTES == 8
-typedef int64_t mit_WORD;
-typedef uint64_t mit_UWORD;
+#define mit_WORD_BIT 64
 #define mit_SIZE_WORD 3
-#define mit_WORD_MASK UINT64_MAX
-#define mit_UWORD_MAX UINT64_MAX
-#define mit_WORD_MIN INT64_MIN
-#define mit_WORD_MAX INT64_MAX
-#define reverse_endianness bswap_64
 #elif WORD_BYTES == 16
+#define mit_WORD_BIT 128
+#define mit_SIZE_WORD 4
 /* Given an unsigned 128-bit argument X, return the value corresponding to
    X with reversed byte order.  */
-#define bswap_128(x) ((((x) & (((unsigned __int128)0xFF << 0)))   << 120) | \
-                      (((x) & (((unsigned __int128)0xFF << 8)))   << 104) | \
-                      (((x) & (((unsigned __int128)0xFF << 16)))  << 88) | \
-                      (((x) & (((unsigned __int128)0xFF << 24)))  << 72) | \
-                      (((x) & (((unsigned __int128)0xFF << 32)))  << 56) | \
-                      (((x) & (((unsigned __int128)0xFF << 40)))  << 40) | \
-                      (((x) & (((unsigned __int128)0xFF << 48)))  << 24) | \
-                      (((x) & (((unsigned __int128)0xFF << 56)))  << 8) | \
-                      (((x) & (((unsigned __int128)0xFF << 64)))  >> 8) | \
-                      (((x) & (((unsigned __int128)0xFF << 72)))  >> 24) | \
-                      (((x) & (((unsigned __int128)0xFF << 80)))  >> 40) | \
-                      (((x) & (((unsigned __int128)0xFF << 88)))  >> 56) | \
-                      (((x) & (((unsigned __int128)0xFF << 96)))  >> 72) | \
-                      (((x) & (((unsigned __int128)0xFF << 104))) >> 88) | \
-                      (((x) & (((unsigned __int128)0xFF << 112))) >> 104) | \
-                      (((x) & (((unsigned __int128)0xFF << 120))) >> 120))
+#define bswap_128(x)                                                    \
+    (((unsigned __int128)(bswap_64((uint64_t)(x))) << 64) |             \
+     ((unsigned __int128)(bswap_64((uint64_t)((x) >> 64)))))
 
 typedef __int128 mit_WORD;
 typedef unsigned __int128 mit_UWORD;
-#define mit_SIZE_WORD 4
-#define mit_WORD_MASK ((__int128)-1)
-#define mit_UWORD_MAX mit_WORD_MASK
-#define mit_WORD_MIN (((unsigned __int128)1) << 127)
-#define mit_WORD_MAX ((((__int128)0x7fffffffffffffff) << 64) | 0xffffffffffffffff)
-#define reverse_endianness bswap_128
 #else
 #error "WORD_BYTES must be 2, 4, 8 or 16!"
 #endif
 
+typedef uint8_t mit_BYTE;
+#define mit_BYTE_BIT 8
+#define mit_BYTE_MASK ((1 << mit_BYTE_BIT) - 1)
+
+#define mit_WORD_MASK ((mit_UWORD)((mit_WORD)-1))
+#define mit_UWORD_MAX mit_WORD_MASK
+#define mit_WORD_MIN ((mit_WORD)(mit_UWORD_MAX << (mit_WORD_BIT - 1)))
+#define mit_WORD_MAX ((mit_WORD)(mit_UWORD_MAX >> 1))
+#define _reverse_endianness(bits, n) bswap_ ## bits(n)
+#define reverse_endianness(bits, n) _reverse_endianness(bits, n)
+
+#if WORD_BYTES != 16
+#define _bits_type_name(prefix, bits) prefix ## bits ## _t
+#define bits_type_name(prefix, bits) _bits_type_name(prefix, bits)
+typedef bits_type_name(int, mit_WORD_BIT) mit_WORD;
+typedef bits_type_name(uint, mit_WORD_BIT) mit_UWORD;
+#endif
 #if SIZEOF_SIZE_T > WORD_BYTES
 typedef size_t mit_max_stack_item_t;
 #else
 typedef mit_UWORD mit_max_stack_item_t;
 #endif
 
-// Constants
+// Constants (for easy FFI binding)
 extern const unsigned mit_word_bytes;
 extern const unsigned mit_endism;
 extern const unsigned mit_size_word;
-#define mit_BYTE_BIT 8
 extern const unsigned mit_byte_bit;
 extern const unsigned mit_byte_mask;
-#define mit_BYTE_MASK ((1 << mit_BYTE_BIT) - 1)
 extern const unsigned mit_word_bit;
-#define mit_WORD_BIT (WORD_BYTES * mit_BYTE_BIT)
 extern const mit_UWORD mit_word_mask;
 extern const mit_UWORD mit_uword_max;
 extern const mit_WORD mit_word_min;
