+++ src/mit_core/instruction.py
@@ -1,7 +1,7 @@
 '''
 Class for instructions.
 
-Copyright (c) 2009-2019 Mit authors
+Copyright (c) 2019 Mit authors
 
 The package is distributed under the MIT/X11 License.
 
@@ -49,16 +49,3 @@ class AbstractInstruction(Enum):
                 instruction.opcode,
             ))
         print('};')
-
-def disable_warnings(warnings, c_source):
-    '''
-    Returns `c_source` wrapped in "#pragmas" to suppress the given list
-    `warnings` of warning flags.
-    '''
-    return '''\
-#pragma GCC diagnostic push
-{pragmas}
-{c_source}
-#pragma GCC diagnostic pop'''.format(c_source=c_source,
-                                     pragmas='\n'.join(['#pragma GCC diagnostic ignored "{}"'.format(w)
-                                                        for w in warnings]))
